pyproject.toml
[project]
name = "smash-fix"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "cloudinary~=1.44.1",
    "django~=5.2.4",
    "django-cloudinary-storage>=0.3.0",
    "djangorestframework~=3.16.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv~=1.1.1",
    "requests~=2.32.4",
]
